common/pyproject.toml
[project]
name = "common"
version = "0.5.0"
description = "Common package"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opentelemetry-exporter-otlp>=1.33.0",
    "pydantic>=2.11.4",
    "pyroscope-otel>=0.4.1",
    "sentry-sdk>=2.27.0",
    "structlog>=25.3.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.8",
]

[tool.ruff]
line-length = 120
target-version="py313"
src = ["src"]
include = ["src/**.py"]

lint.select = [
    "ALL",
]

lint.ignore = [
    "ARG",
    "ANN",
    "D",
    "EM101",
    "EM102",
    "PT001",
    "PT023",
    "SIM108",
    "SIM114",
    "TRY003",
    "PLW2901",
    "RET505",
    "ISC001",
    "UP035",
    "COM812",
    "PTH123",
    "G004",
]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.isort]
known-first-party = ["src"]
no-lines-before = ["local-folder"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88

[tool.mypy]
packages = ["src"]
python_version = "3.13"

strict = true
pretty = true
allow_redefinition = true
allow_untyped_calls = true
disallow_any_explicit = false
disallow_any_generics = false
disallow_subclassing_any = false
disallow_untyped_decorators = false
ignore_missing_imports = true
show_column_numbers = true
show_error_context = true
warn_unreachable = true
